requirements.txt
apache-airflow-providers-amazon
apache-airflow-providers-snowflake
boto3